requirements.txt
asgiref==3.7.2
Babel==2.12.1
crispy-bootstrap5==0.7
da-vinci==0.3.0
Django==4.2.2
django-ajax-datatable==4.4.5
django-appconf==1.0.5
django-bootstrap-modal-forms==3.0.4
django-cleanup==8.0.0
django-crispy-forms==2.0
django-htmx==1.15.0
django-image-cropping==1.7
django-multi-form-view==2.0.1
django-quill-editor==0.1.40
django-select2==8.1.2
django-thumbnails==0.7.0
django-user-agents==0.4.0
easy-thumbnails==2.8.5
pendulum==2.1.2
Pillow==10.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pytzdata==2020.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
ua-parser==0.16.1
user-agents==2.2.0
